+++ libhesai/UdpProtocol/udp_protocol_header.h
@@ -93,7 +93,7 @@ struct HS_LIDAR_PRE_HEADER {
   bool IsValidDelimiter() const {
     return little_to_native(m_u16Delimiter) == kDelimiter;
   }
-  uint8_t GetDelimiter() const { return little_to_native(m_u16Delimiter); }
+  uint16_t GetDelimiter() const { return little_to_native(m_u16Delimiter); }
   uint8_t GetVersionMajor() const { return m_u8VersionMajor; }
   uint8_t GetVersionMinor() const { return m_u8VersionMinor; }
   uint8_t GetStatusInfoVersion() const { return m_u8StatusInfoVersion; }
